+++ CHANGES.md
@@ -1,4 +1,13 @@
 
+# 0.3.0 / 2020-12-04
+
+  * Identify vocabulary lists
+
+# 0.2.0 / 2020-12-03
+
+  * Create sets of unique shingles
+
 # 0.1.0 / 2020-12-02
 
   * Initial Release
+  * Shingling functions